+++ src/davinci_crawling/management/commands/utils/consumer.py
@@ -1,24 +1,20 @@
 # -*- coding: utf-8 -*
 # Copyright (c) 2019 BuildGroup Data Services Inc.
 import logging
-import queue
 import time
 from datetime import datetime
-from threading import Thread
-
 from davinci_crawling.management.commands.utils.utils import \
     update_task_status, get_crawler_by_name
-from davinci_crawling.utils import setup_cassandra_object_mapper
-
-from multiprocessing import Queue
-
-from davinci_crawling.task.models import STATUS_IN_PROGRESS, STATUS_FAULTY,\
+from davinci_crawling.task.models import STATUS_IN_PROGRESS, STATUS_FAULTY, \
     STATUS_FINISHED
+from persistqueue import Queue
+from persistqueue.exceptions import Empty
+from threading import Thread
 
 _logger = logging.getLogger("davinci_crawling.queue")
 
 # stores the multiprocess queue used between crawl_params and crawl method
-multiprocess_queue = Queue()
+multiprocess_queue = Queue("tasks_queue")
 
 
 class CrawlConsumer(object):
@@ -64,11 +60,6 @@ class CrawlConsumer(object):
         Returns:
             The result of the crawler call.
         """
-        # We need to setup the Cassandra Object Mapper to work on
-        # multiprocessing If we do not do that, the processes will be blocked
-        # when interacting with the object mapper module
-        setup_cassandra_object_mapper()
-
         crawler = get_crawler_by_name(crawler_name)
         _logger.debug("Calling crawl method: {0}".
                       format(getattr(crawler, "crawl")))
@@ -95,8 +86,7 @@ class CrawlConsumer(object):
                 return
             task_id = None
             try:
-                crawl_param, options = multiprocess_queue.get(
-                    block=False)
+                crawl_param, options = multiprocess_queue.get(block=False)
                 crawler_name = options.get("crawler")
                 task_id = options.get("task_id")
                 update_task_status(task_id, STATUS_IN_PROGRESS)
@@ -107,7 +97,8 @@ class CrawlConsumer(object):
 
                 self._crawl(crawler_name, task_id, crawl_param, options)
                 update_task_status(task_id, STATUS_FINISHED)
-            except queue.Empty:
+                multiprocess_queue.task_done()
+            except Empty:
                 # Means that the queue is empty and we need to count many times
                 # that the occurs to the close logic, we just start counting
                 # when at least the queue received one
@@ -117,6 +108,7 @@ class CrawlConsumer(object):
             except Exception as e:
                 if task_id:
                     update_task_status(task_id, STATUS_FAULTY)
+                    multiprocess_queue.task_done()
 
                 # TODO add the error message to the db
                 _logger.error("Error while crawling params from queue", e)