+++ src/davinci_crawling/management/commands/utils/utils.py
@@ -9,18 +9,24 @@ from davinci_crawling.task.models import Task
 cached_crawlers = {}
 
 
-def update_task_status(task, status):
+def update_task_status(task, status, more_info=None):
     """
     Get a task td to change the status on the DB.
     Args:
         task: the task that should be updated, can be either a django model or
         a UUID for the task.
         status: the status that we want to put the task on.
+        more_info: more info about the status that is being reported, generally
+        use on the FAULTY state.
     """
     if not isinstance(task, Task):
         task = Task.objects.filter(task_id=task).first()
 
-    task.update(**{"status": status})
+    data = {"status": status}
+    if more_info:
+        data["more_info"] = more_info
+
+    task.update(**data)
 
     return task