+++ src/davinci_crawling/net.py
@@ -9,11 +9,16 @@ import os
 import tempfile
 
 from time import sleep
+
+from davinci_crawling.proxy.proxy import ProxyManager
 from requests.exceptions import RequestException
 from bs4 import BeautifulSoup
 
 from davinci_crawling.exceptions import DownloadException
 from davinci_crawling.io import copy_file
+
+from selenium.common.exceptions import TimeoutException
+from selenium.webdriver.support.wait import WebDriverWait
 
 logger = logging.getLogger("davinci_crawling")
 
@@ -74,10 +79,45 @@ def delete_json(url, timeout=None):
     try:
         timeout = timeout if timeout else DEFAULT_TIMEOUT
 
+        proxy_address = get_proxy_address()
+
         return requests.delete(
             url=url,
             headers={**APPLICATION_JSON, **USER_AGENT},
-            timeout=(timeout, timeout), verify=False)
+            timeout=(timeout, timeout), verify=False, proxies=proxy_address)
+    except RequestException as ex:
+        logger.exception("Unable to send the POST. Cause: %s" % ex)
+        raise ex
+
+
+def get_json(url, timeout=None, custom_header=None, use_proxy=True):
+    """
+    Send a delete request to the given url.
+    Args:
+        url: the url where to call the delete request.
+        timeout: the timeout for the request, if it's none we will use the
+        `DEFAULT_TIMEOUT`
+        custom_header: customize the header with additional headers
+        use_proxy: if true should use the proxy, this is used because before
+        getting the proxies we need to make a get request to the proxy mesh
+        api so we can't use proxy yet
+
+    Returns:
+        the result of the request, object of the requests library
+    """
+    try:
+        timeout = timeout if timeout else DEFAULT_TIMEOUT
+
+        header = {**APPLICATION_JSON, **USER_AGENT}
+        if custom_header:
+            header.update(custom_header)
+
+        proxy_address = get_proxy_address() if use_proxy else {}
+
+        return requests.get(
+            url=url,
+            headers=header,
+            timeout=(timeout, timeout), verify=False, proxies=proxy_address)
     except RequestException as ex:
         logger.exception("Unable to send the POST. Cause: %s" % ex)
         raise ex
@@ -94,10 +134,12 @@ def post_json(url, json_obj, timeout=None):
 
         timeout = timeout if timeout else DEFAULT_TIMEOUT
 
+        proxy_address = get_proxy_address()
+
         return requests.post(
             url=url, data=json_body,
             headers={**APPLICATION_JSON, **USER_AGENT},
-            timeout=(timeout, timeout), verify=False)
+            timeout=(timeout, timeout), verify=False, proxies=proxy_address)
     except RequestException as ex:
         logger.exception("Unable to send the POST. Cause: %s" % ex)
         raise ex
@@ -112,10 +154,12 @@ def post_form(url, json_obj, timeout=None):
     try:
         timeout = timeout if timeout else DEFAULT_TIMEOUT
 
+        proxy_address = get_proxy_address()
+
         return requests.post(
             url=url, data=json_obj,
             headers={**APPLICATION_FORM, **USER_AGENT},
-            timeout=(timeout, timeout), verify=False)
+            timeout=(timeout, timeout), verify=False, proxies=proxy_address)
     except RequestException as ex:
         logger.exception("Unable to send the POST. Cause: %s" % ex)
         raise ex
@@ -126,8 +170,10 @@ def fetch_page(url, timeout):
     Get a raw page fro the url
     """
     logger.info("Fetching page for %s" % url)
+    proxy_address = get_proxy_address()
     return requests.get(url, headers=USER_AGENT,
-                        timeout=(timeout, timeout), verify=False)
+                        timeout=(timeout, timeout), verify=False,
+                        proxies=proxy_address)
 
 
 def parse_json(s):
@@ -160,6 +206,35 @@ def fetch_html(url, timeout=None):
 
 def __constatly_true(status):
     return True
+
+
+def wait_tenaciously(driver, timeout, expected_conditions, n, s,
+                     error_callback=__constatly_true):
+    n -= 1
+    try:
+        # Wait until the page is loaded
+        WebDriverWait(driver, timeout=timeout).until(*expected_conditions)
+    except TimeoutException as ex:
+        logger.exception(
+            f"Exception thrown"
+            f" for {driver.current_url} - ({expected_conditions}): {ex}."
+            f" Max {n} attempts remaining")
+        __maybe_wait(
+            driver, timeout, expected_conditions, n, s, ex, error_callback)
+
+
+def __maybe_wait(driver, timeout, expected_conditions, n, s,
+                 exception, error_callback=None):
+    if error_callback and error_callback(driver, exception) and (n > 0):
+        logger.info("Sleeping for %s seconds before next retry." % s)
+        sleep(s)
+        return wait_tenaciously(
+            driver, timeout, expected_conditions, n, s,
+            error_callback=error_callback)
+
+    logger.info("No more tries (exhausted or "
+                "short-circuited by error callback")
+    raise exception
 
 
 def fetch_tenaciously(fetcher, url, n, s,
@@ -207,8 +282,10 @@ def fetch_file(url, options):
         logger.info(
             "Download from [%s] and store into [%s]" % (url, temp_path))
         try:
+            proxy_address = get_proxy_address()
             response = requests.get(
-                url, stream=True, timeout=(1800, 1800), verify=False)
+                url, stream=True, timeout=(1800, 1800), verify=False,
+                proxies=proxy_address)
 
             params = cgi.parse_header(
                 response.headers.get('Content-Disposition', ''))[-1]
@@ -238,3 +315,12 @@ def fetch_file(url, options):
         except Exception as ex:
             logger.exception("Failed to download %s" % url)
             raise DownloadException(url) from ex
+
+
+def get_proxy_address():
+    proxy_address = ProxyManager().get_proxy_address()
+    if not proxy_address:
+        proxy_address = {}
+    else:
+        proxy_address = proxy_address
+    return proxy_address