+++ src/davinci_crawling/settings.py
@@ -133,6 +133,9 @@ class Common(Configuration):
                                   "/Applications/Chromium.app/Contents"
                                   "/MacOS/Chromium")
 
+    PROXY_MESH_USER = os.getenv("PROXY_MESH_USER")
+    PROXY_MESH_PASSWORD = os.getenv("PROXY_MESH_PASSWORD")
+
     # All the fixed settings that a crawler can have, every crawler should
     # add their specific params here
     DAVINCI_CONF = {
@@ -156,6 +159,18 @@ class Common(Configuration):
             "throttle": {
                 "implementation": "davinci_crawling.throttle.memory_throttle."
                                   "MemoryThrottle"
+            },
+            "proxy": {
+                # to change proxy usage add the implementation key
+                "proxy_mesh": {
+                    "authentication":
+                        "Basic aWFucmlja2V5OlZhcmlhYmxlZGF0YTEwMSE=",
+                    "authorized_proxies_url":
+                        "https://proxymesh.com/api/proxies/"
+                },
+                "proxies-availability-checker": {
+                    "elapse-time-between-checks": 60
+                }
             },
             "parallelism": {
                 "multiproc": {
@@ -443,7 +458,9 @@ class Common(Configuration):
 
         # https://www.django-rest-framework.org/api-guide/fields/#decimalfield
         # To use decimal as representation by default
-        'COERCE_DECIMAL_TO_STRING': False
+        'COERCE_DECIMAL_TO_STRING': False,
+        'EXCEPTION_HANDLER':
+            'caravaggio_rest_api.drf.exceptions.caravaggio_exception_handler'
     }
 
     ACCOUNT_USER_MODEL_USERNAME_FIELD = None
@@ -754,6 +771,8 @@ class Staging(Common):
         }
     }
 
+    THROTTLE_ENABLED = True
+
     # Security
     SECURE_SSL_REDIRECT = os.getenv('SECURE_SSL_REDIRECT', "False") == "True"
     USE_X_FORWARDED_HOST = SECURE_SSL_REDIRECT
@@ -776,3 +795,4 @@ class Production(Staging):
     """
 
     LOGGING_FILE = "/var/log/davinci_crawling-debug.log"
+    THROTTLE_ENABLED = True